added pages to TestNG.xml
